pyproject.toml
[project]
name = "stock-words"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10.0",
    "beautifulsoup4>=4.12.0",
    "fake-useragent>=2.2.0",
    "fastapi>=0.95.0",
    "finance-datareader>=0.9.50",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "matplotlib>=3.7.0",
    "mecab-ko>=1.0.2",
    "mecab-ko-dic>=1.0.0",
    "mecab-python3>=1.0.6",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pika>=1.3.2",
    "playwright>=1.57.0",
    "polars>=0.19.0",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.0",
    "pysocks>=1.7.1",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.21",
    "requests>=2.31.0",
    "schedule>=1.2.2",
    "scikit-learn>=1.3.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.22.0",
    "wordcloud>=1.9.0",
]
